Update Catalog.cpp
Hi guys, 

Awesome project, I was really impressed! You'll see here that I did a few things to your program to just add to it. I have no idea if you ran into this warning, but I received a "deprecated conversion from string constant to ‘char*’" warning when compiling. This was NOT an issue/error as the code still ran totally fine, but I decided to go investigate the cause. I figured out that for some reason your code was having an internal battle between C++ syntax and C syntax, so that warning is given to say that the char* might not work right (but in your case it was fine). So, I went in and corrected the conflicting syntax to all agree and the warning went away! If you see "const char" instead of just "char" throughout your code, that is why.
Again, not sure if that was just my end's issue, so do with that info as you please.

Also, I know this nifty little trick to change font colors, so I made your category types print out in blue to help organize things on the screen a little better. Hopefully you like it but if not no worries!

Again, great job. I really enjoyed using and messing around with all your program can do!
Cheers.
-Matt